+++ views/layouts/main.handlebars
@@ -14,7 +14,7 @@
 
 <body height="100%">
   <header class="d-flex justify-content-between align-items-center p-3">
-    <img class="logo rounded-circle animate__animated animate__bounceInRight" src="/images/LogoImage.png" alt="In-Or-Out Logo" height="125px">
+    <a href="/"> <img class="logo rounded-circle animate__animated animate__bounceInRight" src="/images/LogoImage.png" alt="In-Or-Out Logo" height="125px"></a>
     {{!-- Remove title --}}
     {{!-- <h1>In-Or-Out</h1>  --}}
