+++ views/partials/eventCards.handlebars
@@ -1,4 +1,4 @@
-
+{{!-- todo add grid container back in --}}
 <div class="card animate__animated animate__flipInY">
     <h2 class="title">{{title}}</h2>
     <p class="creator">Created by: {{created_by}}</p>
